requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Deprecated==1.2.9
idna==2.9
PyGithub==1.50
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.9
wrapt==1.12.1
